demo commit

+++ app/src/main/java/com/stu/oyy/record/AudioRecorder.java
@@ -5,9 +5,10 @@ import android.media.AudioRecord;
 import android.media.MediaRecorder;
 import android.text.TextUtils;
 import android.util.Log;
+import com.stu.oyy.util.Async;
+import com.stu.oyy.util.FileUtil;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -70,9 +71,11 @@ public class AudioRecorder {
     public void createAudio(String fileName, int audioSource, int sampleRateInHz, int channelConfig, int audioFormat) {
         // 获得缓冲区字节大小
         bufferSizeInBytes = AudioRecord.getMinBufferSize(sampleRateInHz,
-                channelConfig, channelConfig);
+                channelConfig, audioFormat);
+
         audioRecord = new AudioRecord(audioSource, sampleRateInHz, channelConfig, audioFormat, bufferSizeInBytes);
         this.fileName = fileName;
+        status = Status.STATUS_READY;
     }
 
     /**
@@ -81,12 +84,7 @@ public class AudioRecorder {
      * @param fileName 文件名
      */
     public void createDefaultAudio(String fileName) {
-        // 获得缓冲区字节大小
-        bufferSizeInBytes = AudioRecord.getMinBufferSize(AUDIO_SAMPLE_RATE,
-                AUDIO_CHANNEL, AUDIO_ENCODING);
-        audioRecord = new AudioRecord(AUDIO_INPUT, AUDIO_SAMPLE_RATE, AUDIO_CHANNEL, AUDIO_ENCODING, bufferSizeInBytes);
-        this.fileName = fileName;
-        status = Status.STATUS_READY;
+        createAudio(fileName, AUDIO_INPUT, AUDIO_SAMPLE_RATE, AUDIO_CHANNEL, AUDIO_ENCODING);
     }
 
 
@@ -96,29 +94,20 @@ public class AudioRecorder {
      * @param listener 音频流的监听
      */
     public void startRecord(final RecordStreamListener listener) {
-
-        if (status == Status.STATUS_NO_READY || TextUtils.isEmpty(fileName)) {
+        if (status == Status.STATUS_NO_READY || TextUtils.isEmpty(fileName))
             throw new IllegalStateException("录音尚未初始化,请检查是否禁止了录音权限~");
-        }
-        if (status == Status.STATUS_START) {
+        if (status == Status.STATUS_START)
             throw new IllegalStateException("正在录音");
-        }
-        Log.d("AudioRecorder", "===startRecord===" + audioRecord.getState());
+
         audioRecord.startRecording();
 
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                writeDataTOFile(listener);
-            }
-        }).start();
+        Async.run(() -> writeDataTOFile(listener));
     }
 
     /**
      * 暂停录音
      */
     public void pauseRecord() {
-        Log.d("AudioRecorder", "===pauseRecord===");
         if (status != Status.STATUS_START) {
             throw new IllegalStateException("没有在录音");
         } else {
@@ -130,44 +119,40 @@ public class AudioRecorder {
     /**
      * 停止录音
      */
-    public void stopRecord() {
-        Log.d("AudioRecorder", "===stopRecord===");
-        if (status == Status.STATUS_NO_READY || status == Status.STATUS_READY) {
+    public WaveFile stopRecord() {
+        if (status == Status.STATUS_NO_READY || status == Status.STATUS_READY)
             throw new IllegalStateException("录音尚未开始");
-        } else {
-            audioRecord.stop();
-            status = Status.STATUS_STOP;
-            release();
-        }
+
+        audioRecord.stop();
+        status = Status.STATUS_STOP;
+        return release();
     }
 
     /**
      * 释放资源
      */
-    public void release() {
-        Log.d("AudioRecorder", "===release===");
-        //假如有暂停录音
-        try {
-            if (filesName.size() > 0) {
-                List<String> filePaths = new ArrayList<>();
-                for (String fileName : filesName) {
-                    filePaths.add(FileUtil.getPcmFileAbsolutePath(fileName));
-                }
-                //清除
-                filesName.clear();
-                //将多个pcm文件转化为wav文件
-                mergePCMFilesToWAVFile(filePaths);
+    public WaveFile release() {
+        WaveFile result = null;
 
-            } else {
-                //这里由于只要录音过filesName.size都会大于0,没录音时fileName为null
-                //会报空指针 NullPointerException
-                // 将单个pcm文件转化为wav文件
-                //Log.d("AudioRecorder", "=====makePCMFileToWAVFile======");
-                //makePCMFileToWAVFile();
+        //假如有暂停录音
+        if (filesName.size() > 0) {
+            List<String> filePaths = new ArrayList<>();
+            for (String fileName : filesName) {
+                filePaths.add(FileUtil.getPcmFileAbsolutePath(fileName));
             }
-        } catch (IllegalStateException e) {
-            throw new IllegalStateException(e.getMessage());
+            //清除
+            filesName.clear();
+            //将多个pcm文件转化为wav文件
+            result = mergePCMFilesToWAVFile(filePaths);
+
+        } else {
+            //这里由于只要录音过filesName.size都会大于0,没录音时fileName为null
+            //会报空指针 NullPointerException
+            // 将单个pcm文件转化为wav文件
+            //Log.d("AudioRecorder", "=====makePCMFileToWAVFile======");
+            //makePCMFileToWAVFile();
         }
+
 
         if (audioRecord != null) {
             audioRecord.release();
@@ -175,6 +160,7 @@ public class AudioRecorder {
         }
 
         status = Status.STATUS_NO_READY;
+        return result;
     }
 
     /**
@@ -198,53 +184,46 @@ public class AudioRecorder {
      * @param listener 音频流的监听
      */
     private void writeDataTOFile(RecordStreamListener listener) {
-        // new一个byte数组用来存一些字节数据，大小为缓冲区大小
-        byte[] audiodata = new byte[bufferSizeInBytes];
-
         FileOutputStream fos = null;
-        int readsize = 0;
         try {
+            // new一个byte数组用来存一些字节数据，大小为缓冲区大小
+            byte[] audiodata = new byte[bufferSizeInBytes];
+            int readsize = 0;
             String currentFileName = fileName;
             if (status == Status.STATUS_PAUSE) {
                 //假如是暂停录音 将文件名后面加个数字,防止重名文件内容被覆盖
                 currentFileName += filesName.size();
-
             }
+
             filesName.add(currentFileName);
             File file = new File(FileUtil.getPcmFileAbsolutePath(currentFileName));
-            if (file.exists()) {
-                file.delete();
-            }
+            if (file.exists()) file.delete();
             fos = new FileOutputStream(file);// 建立一个可存取字节的文件
-        } catch (IllegalStateException e) {
-            Log.e("AudioRecorder", e.getMessage());
-            throw new IllegalStateException(e.getMessage());
-        } catch (FileNotFoundException e) {
-            Log.e("AudioRecorder", e.getMessage());
 
-        }
-        //将录音状态设置成正在录音状态
-        status = Status.STATUS_START;
-        while (status == Status.STATUS_START) {
-            readsize = audioRecord.read(audiodata, 0, bufferSizeInBytes);
-            if (AudioRecord.ERROR_INVALID_OPERATION != readsize && fos != null) {
-                try {
+            //将录音状态设置成正在录音状态
+            status = Status.STATUS_START;
+            while (status == Status.STATUS_START) {
+                readsize = audioRecord.read(audiodata, 0, bufferSizeInBytes);
+                if (AudioRecord.ERROR_INVALID_OPERATION != readsize) {
                     fos.write(audiodata);
+
+                    //用于拓展业务
                     if (listener != null) {
-                        //用于拓展业务
                         listener.recordOfByte(audiodata, 0, audiodata.length);
                     }
-                } catch (IOException e) {
-                    Log.e("AudioRecorder", e.getMessage());
                 }
             }
-        }
-        try {
-            if (fos != null) {
-                fos.close();// 关闭写入流
-            }
+
         } catch (IOException e) {
             Log.e("AudioRecorder", e.getMessage());
+        } finally {
+            if (fos != null) {
+                try {
+                    fos.close();// 关闭写入流
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
         }
     }
 
@@ -253,39 +232,37 @@ public class AudioRecorder {
      *
      * @param filePaths
      */
-    private void mergePCMFilesToWAVFile(final List<String> filePaths) {
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                if (PcmToWav.mergePCMFilesToWAVFile(filePaths, FileUtil.getWavFileAbsolutePath(fileName))) {
-                    //操作成功
-                } else {
-                    //操作失败
-                    Log.e("AudioRecorder", "mergePCMFilesToWAVFile fail");
-                    throw new IllegalStateException("mergePCMFilesToWAVFile fail");
-                }
-                fileName = null;
-            }
-        }).start();
+    private WaveFile mergePCMFilesToWAVFile(final List<String> filePaths) {
+
+        WaveFile waveFile = PcmToWav.mergePCMFilesToWAVFile(filePaths);
+
+        if (waveFile == null)
+            throw new IllegalStateException("fail to convert to WavFile Object!!");
+
+//        if (!waveFile.toFile(FileUtil.getWavFileAbsolutePath(fileName)))
+//            throw new IllegalStateException("WavFile Object fail to convert to file!!");
+
+        fileName = null;
+
+        return waveFile;
+
     }
 
     /**
      * 将单个pcm文件转化为wav文件
      */
     private void makePCMFileToWAVFile() {
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                if (PcmToWav.makePCMFileToWAVFile(FileUtil.getPcmFileAbsolutePath(fileName), FileUtil.getWavFileAbsolutePath(fileName), true)) {
-                    //操作成功
-                } else {
-                    //操作失败
-                    Log.e("AudioRecorder", "makePCMFileToWAVFile fail");
-                    throw new IllegalStateException("makePCMFileToWAVFile fail");
-                }
-                fileName = null;
-            }
-        }).start();
+
+        WaveFile waveFile = PcmToWav.makePCMFileToWAVFile(FileUtil.getPcmFileAbsolutePath(fileName), true);
+
+        if (waveFile == null)
+            throw new IllegalStateException("fail to convert to WavFile Object!!");
+
+//        if (!waveFile.toFile(FileUtil.getWavFileAbsolutePath(fileName)))
+//            throw new IllegalStateException("WavFile Object fail to convert to file!!");
+
+        fileName = null;
+
     }
 
     /**
